requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.6.0
django-elasticsearch-dsl==7.1.4
django-oauth-toolkit==1.3.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
gunicorn==20.0.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2-binary==2.8.6
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.2
wrapt==1.12.1
